Updated functions
removed GameNodeData class, now data is stored seperately in the TreeNode class. Ponts still dont work correctly but here's tips from a random person on the stardew valley discord:

so there's a few ways to fix this. the easiest ways are
pass the depth as a parameter to GenerateTree
pass the parent node as a parameter to GenerateTree and have the child add itself
unwrap the recursion into a queue/stack (ie. make it a loop)